--- src/main/java/unicon/Achiva/domain/article/infrastructure/CounterHelper.java
@@ -1,37 +1,0 @@
-package unicon.Achiva.domain.article.infrastructure;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import unicon.Achiva.domain.category.Category;
-import unicon.Achiva.domain.member.entity.MemberCategoryCounter;
-import unicon.Achiva.domain.member.entity.MemberCategoryKey;
-import unicon.Achiva.domain.member.infrastructure.MemberCategoryCounterRepository;
-
-import java.util.List;
-import java.util.UUID;
-
-@Service
-@RequiredArgsConstructor
-public class CounterHelper {
-    private final MemberCategoryCounterRepository counterRepo;
-
-    @Transactional
-    public MemberCategoryCounter lockOrInit(UUID memberId, Category category) {
-        MemberCategoryKey key = new MemberCategoryKey(memberId, category);
-        return counterRepo.lockById(key).orElseGet(() -> {
-            MemberCategoryCounter c = new MemberCategoryCounter();
-            c.setId(key);
-            c.setSize(0L);
-            return counterRepo.saveAndFlush(c);
-        });
-    }
-
-    // 교착 방지용: 항상 작은 쪽 먼저 락
-    public List<MemberCategoryKey> orderedKeys(UUID memberId, Category a, Category b) {
-        Category first = a.name().compareTo(b.name()) <= 0 ? a : b;
-        Category second = (first == a) ? b : a;
-        return List.of(new MemberCategoryKey(memberId, first),
-                new MemberCategoryKey(memberId, second));
-    }
-}
